pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "json_agent"
version = "0.1.0"
authors = [
  { name="Your Name", email="your@email.com" },
]
description = "An agent that can read and modify JSON files based on a prompt."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.34.1",
    "python-multipart==0.0.9",
    "python-dotenv==1.0.1",
]

[project.scripts]
json-agent-server = "json_agent.__main__:main"
